refac: Scope aws credentials to object and change endpoint_url -> region
Changelog: None
Ticket: None
